+++ electrum/gui/qml/qechannellistmodel.py
@@ -7,8 +7,9 @@ from electrum.logging import get_logger
 from electrum.util import Satoshis
 from electrum.gui import messages
 
+from electrum.gui.common_qt.util import qt_event_listener, QtEventListener
+
 from .qetypes import QEAmount
-from .util import QtEventListener, qt_event_listener
 from .qemodelfilter import QEFilterProxyModel
