+++ electrum/gui/qml/qeinvoicelistmodel.py
@@ -8,7 +8,9 @@ from electrum.logging import get_logger
 from electrum.util import Satoshis, format_time
 from electrum.invoices import BaseInvoice, PR_EXPIRED, LN_EXPIRY_NEVER, Invoice, Request, PR_PAID
 
-from .util import QtEventListener, qt_event_listener, status_update_timer_interval
+from electrum.gui.common_qt.util import QtEventListener, qt_event_listener
+
+from .util import status_update_timer_interval
 from .qetypes import QEAmount
 
 if TYPE_CHECKING:
