+++ electrum/gui/qt/network_dialog.py
@@ -41,11 +41,10 @@ from electrum.logging import get_logger
 from electrum.util import is_valid_websocket_url
 from electrum.gui import messages
 
+from electrum.gui.common_qt.util import QtEventListener, qt_event_listener
 from .util import (
-    Buttons, CloseButton, HelpButton, read_QIcon, char_width_in_lineedit, PasswordLineEdit, QtEventListener,
-    qt_event_listener, Spinner, HelpLabel
+    Buttons, CloseButton, HelpButton, read_QIcon, char_width_in_lineedit, PasswordLineEdit, Spinner, HelpLabel
 )
-
 
 _logger = get_logger(__name__)
